+++ api/api_test.go
@@ -1,4 +1,4 @@
-package handlers
+package api
 
 import (
 	"os"
